Scale model to d=768/8-layer/mpnet and add performance optimizations
Architecture: d_model 384→768, heads 8→12, layers 4→8, ff 1536→3072,
CNN [32,64]→[64,128], text encoder MiniLM→all-mpnet-base-v2 (768-dim).
Result: 122.4M trainable params, 231.9M at inference.

Performance optimizations (no math/model changes):
- GPU-side text embedding cache (eliminates CPU→GPU transfers)
- Split gate linear to avoid 400MB temporary concat tensor
- Cache param lists before training loop
- optimizer.zero_grad(set_to_none=True)
- Single named_parameters() walk for grad norms
- .reshape() instead of .contiguous().view() in attention
- Remove redundant .float() in preprocessing
- num_workers=4 for unseen eval DataLoader

Batch size tuned via full-pipeline benchmarking: BS=64, accum=7
(effective=448). Peak ~15GB worst-case, safe on 25GB RTX 4090.

+++ model/preprocessing.py
@@ -60,7 +60,7 @@ def create_patches(
     patches = patches.squeeze(0)  # (channels, num_patches, patch_timesteps)
     patches = patches.permute(1, 2, 0)  # (num_patches, patch_timesteps, channels)
 
-    return patches.contiguous().float()
+    return patches.contiguous()
 
 
 def interpolate_patches(
@@ -115,7 +115,7 @@ def interpolate_patches(
         x = F.interpolate(x, size=target_size, mode='linear', align_corners=False)
 
     # Back to (num_patches, target_size, num_channels)
-    return x.permute(0, 2, 1).contiguous().float()
+    return x.permute(0, 2, 1).contiguous()
 
 
 def normalize_patches(
